test_patch_upstream_status: refactor using a new regex and removing presence unit test
Individual tests are now combined into a single one that checks both presence and
format with different output messages in case of failure. This reduces the code
and redundancy on checking each line of the payload.

Regarding the regex, current version of the test case is comparing a literal string to
a given payload, meaning it is possible for the unit tests to give false-positives
when there are no matches as current logic expects at least one.
Having this regex will make the presence unit test to FAIL despite Upstream-Status
being present with a bad format. Something similar happens with the format
unit test, which will always PASS as the condition for failure is not satisfied if
bad formatted because there's no match.

This change sets the regex with IGNORECASE flag and also ignores the dash
presence/absence so matches can be used properly; the literal string is compared
when parse_upstream_status.py is called, no need to do such check along the test case.

[YOCTO #11055]

Signed-off-by: Daniela Plascencia <daniela.plascencia@linux.intel.com>